Update test to show on issue in code
This commit updates the test for S3FilesystemV1 to point out a issue in how the code works when trying to access a url for a S3 object using the Filesystem adapter in Laravel.

I have code like this
Storage::disk('public')->url($filename);

where the disk public is configured with:
'driver' => 'async-aws-s3',

The issue that I see is that \AsyncAws\Illuminate\Filesystem\AsyncAwsFilesystemAdapter::url is accessing the S3FilesystemV1 and trying to call getClient on it which does not exists in the version required by this (0.3.3).
It is added and tagged in 0.4.0.