Add support for generating code for discovered endpoint using the URI
The core package has been updated to support those, so the code
generator can rely on them.
This makes the RequirementRegistry smarter by merging requirements for
the same package added in multiple places by keeping the requirement
with the highest lower bound. For now, this supports only the `*`
requirement and requirements using the `^` operator as they are the only
requirements used by the code generators.