Make the code generator manage the async-aws/core requirement
As the registry now supports properly combining multiple min versions
for a semver requirement, we can make the code generator manage the
async-aws/core requirement of generated services even when they don't
require special features instead of relying on the template to add it.

The generators now also avoid adding a requirement on async-aws/core if
the generator code is inside the Core package to avoid generating a
self-referencing requirement (this was fine before only because the Sts
client was not relying on the special features that were adding such a
requirement).

This also removes the json extension from the template as it is managed
by the code generator anyway and makes the symfony/polyfill-uuid
requirement cleanable in case a service does not need it anymore.