Refactor the entrypoint of the runner
The generate.php script in the code generator does not make sense as an
executable as it defaults to reading the manifest and generating the
code in the code generator folder itself, which is not the actual use
case.
Configuring it through environment variables also makes it harder to
configure objects like the directory resolver.