Allow using the patched dev-master branch of core in lowest jobs (#1520)
Thanks to marking the path repository non-canonical for the lowest job,
it will only be used when we don't have a stable release matching the
constraint. But it will be used when the requirement enforces the usage
of the dev version instead of using the version from the existing
subtree split which does not have the in-progress patch of the dev
version.