Fix the local test setup (#1959)
When installing dependencies in the root project, the bundle tests are
installing phpunit as a transitive dependency. This causes issue when
the installed version is incompatible with the version installed by the
simple-phpunit script.
CI was not impacted because it injects such conflict rule in
component-level files to prevent this issue already.